Revert "fix: Fix remove not working on broken symlinks."
This reverts commit 8da440c32df5cd667339d44ed08cbf68d54d8c33.

+++ crates/utils/src/fs.rs
@@ -488,10 +488,12 @@ pub fn read_file_bytes<T: AsRef<Path>>(path: T) -> Result<Vec<u8>, FsError> {
 pub fn remove<T: AsRef<Path>>(path: T) -> Result<(), FsError> {
     let path = path.as_ref();
 
-    if path.is_dir() {
-        remove_dir_all(path)?;
-    } else {
-        remove_file(path)?;
+    if path.exists() {
+        if path.is_file() {
+            remove_file(path)?;
+        } else if path.is_dir() {
+            remove_dir_all(path)?;
+        }
     }
 
     Ok(())
@@ -502,9 +504,7 @@ pub fn remove<T: AsRef<Path>>(path: T) -> Result<(), FsError> {
 pub fn remove_file<T: AsRef<Path>>(path: T) -> Result<(), FsError> {
     let path = path.as_ref();
 
-    // Use `try_exists` so that broken symbolic links still return true
-    // https://doc.rust-lang.org/std/fs/fn.try_exists.html
-    if path.try_exists().is_ok() {
+    if path.exists() {
         trace!(file = ?path, "Removing file");
 
         fs::remove_file(path).map_err(|error| FsError::Remove {
@@ -525,7 +525,7 @@ pub fn remove_file_if_older_than<T: AsRef<Path>>(
 ) -> Result<u64, FsError> {
     let path = path.as_ref();
 
-    if path.try_exists().is_ok() {
+    if path.exists() {
         if let Ok(meta) = metadata(path) {
             let now = SystemTime::now();
             let last_used = meta
@@ -551,7 +551,7 @@ pub fn remove_file_if_older_than<T: AsRef<Path>>(
 pub fn remove_dir_all<T: AsRef<Path>>(path: T) -> Result<(), FsError> {
     let path = path.as_ref();
 
-    if path.try_exists().is_ok() {
+    if path.exists() {
         trace!(dir = ?path, "Removing directory");
 
         fs::remove_dir_all(path).map_err(|error| FsError::Remove {
